refactor: extract formatActivity and add tests
