+++ src/app/api/metrics/activity/route.ts
@@ -1,3 +1,4 @@
+import { formatActivity } from "@/lib/activity-formatter";
 import { getServerSession } from "next-auth";
 import { NextRequest } from "next/server";
 import { authOptions } from "@/lib/auth";
@@ -78,87 +79,6 @@ function capitalize(value: string): string {
   return value.length > 0
     ? value[0].toUpperCase() + value.slice(1)
     : "Updated";
-}
-
-function formatActivity(event: RawEvent): ActivityItem | null {
-  const repoName = event.repo?.name;
-
-  if (!repoName || !SUPPORTED_EVENT_TYPES.has(event.type)) {
-    return null;
-  }
-
-  if (event.type === "PushEvent") {
-    const commitCount = event.payload?.commits?.length ?? 0;
-    const rawRef = event.payload?.ref ?? "";
-    const branch = rawRef.replace("refs/heads/", "") || "default branch";
-    const plural = commitCount === 1 ? "" : "s";
-
-    return {
-      id: event.id,
-      type: "push",
-      createdAt: event.created_at,
-      title: `Pushed ${commitCount} commit${plural} to ${branch}`,
-      subtitle: repoName,
-      repo: repoName,
-      url: event.payload?.head
-        ? `https://github.com/${repoName}/commit/${event.payload.head}`
-        : getRepoUrl(repoName),
-    };
-  }
-
-  if (event.type === "PullRequestEvent") {
-    const action = event.payload?.action ?? "updated";
-    const pr = event.payload?.pull_request;
-    const number = pr?.number ? `#${pr.number}` : "PR";
-    const wasMerged = action === "closed" && pr?.merged === true;
-    const actionText = wasMerged ? "Merged" : capitalize(action);
-
-    return {
-      id: event.id,
-      type: "pull_request",
-      createdAt: event.created_at,
-      title: `${actionText} pull request ${number}`,
-      subtitle: pr?.title ?? repoName,
-      repo: repoName,
-      url: pr?.html_url ?? getRepoUrl(repoName),
-    };
-  }
-
-  if (event.type === "IssuesEvent") {
-    const action = event.payload?.action ?? "updated";
-    const issue = event.payload?.issue;
-    const number = issue?.number ? `#${issue.number}` : "Issue";
-    const actionText = capitalize(action);
-
-    return {
-      id: event.id,
-      type: "issue",
-      createdAt: event.created_at,
-      title: `${actionText} issue ${number}`,
-      subtitle: issue?.title ?? repoName,
-      repo: repoName,
-      url: issue?.html_url ?? getRepoUrl(repoName),
-    };
-  }
-
-  if (event.type === "ReleaseEvent") {
-    const action = event.payload?.action ?? "published";
-    const release = event.payload?.release;
-    const tag = release?.tag_name ?? "release";
-    const actionText = capitalize(action);
-
-    return {
-      id: event.id,
-      type: "release",
-      createdAt: event.created_at,
-      title: `${actionText} ${tag}`,
-      subtitle: release?.name ?? repoName,
-      repo: repoName,
-      url: release?.html_url ?? getRepoUrl(repoName),
-    };
-  }
-
-  return null;
 }
 
 async function fetchFormattedActivity(token: string): Promise<ActivityItem[]> {
